refactor: move regenerate from agent-playbooks to user-playbooks
The regenerate command re-extracts user playbooks from interactions,
not agent playbooks. Having it under agent-playbooks was misleading.

- Moved regenerate command to user_playbooks.py
- Includes auto-promote PENDING→CURRENT after --wait
- agent-playbooks now only has aggregate (which creates agent playbooks
  from user playbooks)
