+++ reflexio/cli/commands/user_playbooks.py
@@ -1,4 +1,4 @@
-"""User playbook management commands (list, search, add, delete)."""
+"""User playbook management commands (list, search, add, delete, regenerate)."""
 
 from __future__ import annotations
 
@@ -18,7 +18,7 @@ from reflexio.cli.output import (
     print_user_playbooks,
     render,
 )
-from reflexio.cli.state import get_client
+from reflexio.cli.state import get_client, require_agent_version
 
 app = typer.Typer(help="Manage user playbooks.")
 
@@ -324,3 +324,57 @@ def delete_all(
         render(resp, json_mode=True)
     else:
         print_info("All user playbooks deleted")
+
+
+@app.command()
+@handle_errors
+def regenerate(
+    ctx: typer.Context,
+    wait: Annotated[
+        bool,
+        typer.Option("--wait", help="Wait for regeneration to complete"),
+    ] = False,
+    agent_version: Annotated[
+        str | None,
+        typer.Option("--agent-version", help="Agent version to regenerate playbooks for"),
+    ] = None,
+) -> None:
+    """Re-extract user playbooks from published interactions.
+
+    Re-runs the playbook extraction pipeline over all interactions for the
+    given agent version, producing fresh user playbooks.
+
+    Args:
+        ctx: Typer context with CliState in ctx.obj
+        wait: If True, wait for regeneration to complete
+        agent_version: Agent version to regenerate (required)
+    """
+    client = get_client(ctx)
+    resolved_version = require_agent_version(
+        agent_version, command_hint="user-playbook regeneration"
+    )
+
+    resp = client.rerun_playbook_generation(
+        agent_version=resolved_version,
+        wait_for_response=wait,
+    )
+
+    # When waiting, auto-promote PENDING user playbooks to CURRENT so
+    # they're immediately visible in `list` output.
+    if wait:
+        upgrade_resp = client.upgrade_user_playbooks(
+            agent_version=resolved_version,
+        )
+        promoted = upgrade_resp.user_playbooks_promoted if upgrade_resp else 0
+    else:
+        promoted = 0
+
+    json_mode: bool = ctx.obj.json_mode
+    if json_mode:
+        render(resp, json_mode=True)
+    elif wait:
+        print_info(
+            f"Playbook regeneration complete ({promoted} playbook(s) promoted)"
+        )
+    else:
+        print_info("Playbook regeneration started")
